confer/requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
alipy==1.2.5
altair==5.5.0
annotated-types==0.7.0
anyio==4.7.0
asttokens==2.4.1
async-timeout==5.0.1
attrs==24.2.0
blinker==1.9.0
Brotli
cachetools==5.5.0
captum==0.7.0
certifi==2023.11.17
charset-normalizer==3.2.0
chicken-dinner==0.11.0
clarabel==0.9.0
click==7.1.2
cloudpickle==3.1.0
colorama==0.4.6
comm==0.2.1
contourpy==1.3.0
cvxpy==1.6.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distro==1.9.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.115.6
fastjsonschema==2.20.0
ffmpeg==1.4
filelock==3.12.2
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
grpclib==0.4.7
h11==0.14.0
h2==4.1.0
hdbscan==0.8.40
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
imbalanced-learn==0.12.4
importlib-metadata==7.0.1
importlib_resources==6.4.5
ipykernel==6.29.0
ipython==8.18.1
jedi==0.19.1
Jinja2==3.1.2
jiter==0.8.0
joblib==1.2.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.7
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.9.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-service==2.4.0
mkl_fft==1.3.11
mkl_random
# modal==0.12.0           <-- 제거 (Python 3.9 환경에서 호환 문제)
modAL-python==0.4.2.1
mpl_interactions==0.24.2
mpmath==1.3.0
multidict==6.1.0
narwhals==1.16.0
nbformat==5.10.4
nest-asyncio==1.5.9
networkx==3.1
numba==0.60.0
numpy
openai==0.28.0
opencv-python==4.10.0.84
osqp==0.6.7.post3
packaging==23.2
pandas==1.5.3
parso==0.8.3
patsy==1.0.1
pillow==9.5.0
platformdirs==4.1.0
plotly==5.24.1
prettytable==3.12.0
prompt-toolkit==3.0.43
propcache==0.2.0
protobuf==5.29.1
psutil==5.9.7
psycopg2-binary==2.9.10
pure-eval==0.2.2
pyarrow==18.1.0
pydantic==2.10.3
pydantic_core==2.27.1
pydeck==0.9.1
Pygments==2.17.2
pyparsing==3.2.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.2
qdldl==0.1.7.post5
referencing==0.35.1
requests==2.31.0
rich==13.9.4
rpds-py==0.21.0
scikit-learn==1.5.2
scipy==1.13.1
scs==3.2.7.post2
seaborn==0.13.2
shap==0.46.0
shellingham==1.5.4
sigtools==4.0.1
six==1.16.0
skorch==0.9.0
slicer==0.0.8
smmap==5.0.1
sniffio==1.3.1
SQLAlchemy==2.0.36
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.4
streamlit==1.40.2
sympy==1.13.1
synchronicity==0.9.8
tabulate==0.8.10
tenacity==9.0.0
threadpoolctl==3.1.0
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typer==0.4.0
types-certifi==2021.10.8.3
types-toml==0.10.8.20240310
typing_extensions==4.12.2
tzdata==2023.3
urllib3==1.26.15
watchdog==6.0.0
watchfiles==1.0.3
wcwidth==0.2.13
win-inet-pton==1.1.0
yarl==1.17.1
zipp==3.17.0
